Move tests to shared project
