+++ sonar-dotnet-shared-library/src/test/resources/AbstractRulesDefinitionTest/Rules.json
@@ -21,5 +21,13 @@
   {
     "id": "S1114",
     "parameters": [ ]
+  },
+  {
+    "id": "S1115",
+    "parameters": [ ]
+  },
+  {
+    "id": "S1116",
+    "parameters": [ ]
   }
 ]
