+++ sonar-dotnet-shared-library/src/test/resources/AbstractRulesDefinitionTest/S1115.html
@@ -1,0 +1,2 @@
+Lorem ipsum
+<b>HTML</b>. <br /><br /><br />End
